pyproject.toml
[project]
name = "online-cinema"
version = "0.1.0"
description = ""
authors = [
    {name = "Anastasia",email = "anastasiiya.p@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-env (>=1.2.0,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt (<4.0.0)",
    "pytest-order (>=1.3.0,<2.0.0)",
    "validators (>=0.35.0,<0.36.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "faker (>=37.11.0,<38.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "aiosmtplib (>=5.0.0,<6.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "fastapi-swagger-dark (>=0.0.8,<0.0.9)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "stripe (>=14.1.0,<15.0.0)",
    "flake8-annotations (>=3.2.0,<4.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
markers = [
    "no_seed: skip database seeding for this test",
]
